+++ example/lib/views/map/map.dart
@@ -1,0 +1,10 @@
+library map;
+
+import 'package:flutter/material.dart';
+import 'package:flutter_map/flutter_map.dart';
+import 'package:layrz_models/layrz_models.dart';
+import 'package:layrz_theme/layrz_theme.dart';
+import 'package:layrz_theme_example/store/store.dart';
+import 'package:vxstate/vxstate.dart';
+
+part 'src/layer.dart';